+++ restart_dada.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ "$#" -ne 1 ]; then
+    echo "Illegal number of parameters"
+    exit 1
+fi
+
 rm /scratch.global/rabdill/bulk/${1}/results/*
 rm /scratch.global/rabdill/bulk/${1}/temp/*
 rm /scratch.global/rabdill/bulk/${1}/intermediate/*
